code cleanup
Separate URLUtil from UIUtils to enable browser-based OAuth in CLI mode even when SWT jar is not present.
